examples/todo: extracts schema
So TodoSchema can be re-use for demo purposes
